pyproject.toml
[project]
name = "mtaa-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.14.1",
  "asyncpg>=0.30.0",
  "faker>=37.1.0",
  "fastapi[standard]>=0.115.8",
  "firebase-admin>=6.7.0",
  "pycountry>=24.6.1",
  "pydantic-extra-types>=2.10.3",
  "pydantic-settings>=2.8.1",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.20",
  "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
 "pre-commit",
 "ruff>=0.9.8",
]
